move liveness and readiness to health.py
